requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.2
cssselect==1.2.0
distlib==0.3.8
Django==5.0.2
djangorestframework==3.14.0
docutils==0.20.1
fastapi==0.109.2
filelock==3.13.1
Flask==3.0.2
greenlet==3.0.3
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kivymd==1.2.0
lxml==5.1.0
MarkupSafe==2.1.5
packaging==23.2
parsel==1.8.1
pillow==10.2.0
pipenv==2023.12.1
platformdirs==4.2.0
playwright==1.41.2
Protego==0.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyDispatcher==2.0.7
pyee==11.0.1
Pygments==2.17.2
pyOpenSSL==24.0.0
pypiwin32==223
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
queuelib==1.6.2
requests==2.31.0
requests-file==2.0.0
Scrapy==2.11.1
service-identity==24.1.0
setuptools==69.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
starlette==0.36.3
tldextract==5.1.1
Twisted==23.10.0
twisted-iocpsupport==1.0.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
virtualenv==20.25.1
w3lib==2.1.2
Werkzeug==3.0.1
zope.interface==6.1
